lemur-combined/requirements.txt

Flask==2.0.1
google-auth==2.0.1
google-auth-oauthlib==0.4.6
google-cloud-pubsub==2.7.0
google-api-python-client==2.21.0
gunicorn==20.1.0
requests==2.26.0
werkzeug==2.0.3
pytz==2021.1
google-cloud-logging
google-auth
python-pptx
